TelegramBot/requirements.txt
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.0
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tifffile==2020.12.8
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
typing-extensions==3.7.4.3
urllib3==1.26.2
